requirements.txt
Django==3.2.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0